refactor: Refactor init functions
