+++ Makefile
@@ -62,6 +62,11 @@ SRC_MINIMAP				= $(addprefix $(SRC_MINIMAP_DIR), minimap.c)
 SRC_RAYCAST_DIR			= raycast/
 SRC_RAYCAST				= $(addprefix $(SRC_RAYCAST_DIR), raycast.c)
 
+SRC_INIT_DIR			= init/
+SRC_INIT				= $(addprefix $(SRC_INIT_DIR), init.c \
+														init_state.c \
+														init_info.c)
+
 SRC_MAIN_DIR			= main/
 SRC_MAIN				= $(addprefix $(SRC_MAIN_DIR), main.c)
 
@@ -73,7 +78,8 @@ SRC						= $(SRC_MAIN) \
 							$(SRC_ERROR) \
 							$(SRC_MINIMAP) \
 							$(SRC_UTIL)	\
-							$(SRC_RAYCAST)
+							$(SRC_RAYCAST) \
+							$(SRC_INIT)
 
 SRCS_DIR				= ./src/
 SRCS					= $(addprefix $(SRCS_DIR), $(SRC))
