+++ map/overflow.cub
@@ -1,0 +1,23 @@
+NO ./img/NO.xpm
+SO ./img/SO.xpm
+WE ./img/WE.xpm
+EA ./img/EA.xpm
+F 00,153,15
+C 255,255,255
+
+        1111111111111111111111111111111111111111111111111111111111
+        1000000000000000000000000000000000000000000000000000000001
+        1000000000000000000000000000000000000000000000000000000001
+        1000000000000000000000001111111111111111111111111111111111
+        1000000000000000000000001111111111111111111111111111111111
+        1000000000000000000000001111111111111111111111111111111111
+111111111000000000000000000000001
+100000000000000000000000000001111
+10000000000000E10000000000001
+10000000000000000000000000001
+10000000000000000000000000001
+10000000000000000000000000001
+10000000000000000001010000001
+11000001110101011111011110101
+11110111 1110111 101111010001
+11111111 1111111 111111111111
