+++ src/handler/keypress_handler.c
@@ -52,7 +52,7 @@ static void	handle_minimap(int key, t_game *game)
 	rotate_pos(key, &game->state.dir, &game->state.plane);
 	draw_minimap_bg(game);
 	draw_dir_line(&game->state, game->minimap_info.b_size, &game->gl);
-	draw_minimap_player(game, 3);
+	draw_minimap_player(game);
 	raycast(game, game->minimap_info.b_size * game->map_info.width);
 }
